Fixed transitions on some Android 4.1 devices
On some devices (tested on a Samsung Galaxy S2), Android 4.1 is
erroneously reporting to support unprefixed transitions events, while
it’s in fact not implemented.

The solution is to test for the prefixed version first. The
WebKit-prefixed event has been moved on top of the tested events, and
while the Object properties order is not guaranteed in JavaScript (ES3,
section 4.3.3), it is the case in the targeted browser.