stringToFragment not using cache when templateString != templateString.trim()
The issue was that initially we check the cache by an un-trimmed key, while later on, before putting a value in the cache, we trim the key. This would lead to the fragment of a templateString that has spaces at the beginning or the end, not to be retrieved from the cache.

The fix us to use the original, unchanged value as the key.